Add test for default_escape => 'none'
This parallels what is accepted by 'escape'.

As we are transitioning to escape by default, we have one method that
sets a default of escape = none, and one that sets a default of escape = 'html'.

In our case, we need to override a a setting of default_escape='html'
with a default_escape='none', but this is currently not possible.